requirements.txt

flask
werkzeug
pytesseract
Pillow
fpdf
gunicorn
